feat(fizzbuzz): expose BUZZ_FIZZ_FACTOR_RULES to be reused
